requirements.txt
# gdal
# gdal-utils

boto3
requests
feedparser
python-dateutil

ckanext-scheming~=2.1.0
ckanext-toolbelt~=0.0.7
ckanext-hierarchy~=1.0.2
